+++ packages/tizen_package_manager/example/integration_test/tizen_package_manager_test.dart
@@ -4,7 +4,7 @@
 
 import 'package:flutter_test/flutter_test.dart';
 import 'package:integration_test/integration_test.dart';
-import 'package:tizen_package_manager/package_manager.dart';
+import 'package:tizen_package_manager/tizen_package_manager.dart';
 
 import 'package:tizen_package_manager_example/main.dart';
 
@@ -12,7 +12,6 @@ void main() {
   IntegrationTestWidgetsFlutterBinding.ensureInitialized();
 
   testWidgets('Can get current package info', (WidgetTester tester) async {
-    // These test are based on the example app.
     final PackageInfo info =
         await PackageManager.getPackageInfo(currentPackageId);
     expect(info.packageId, 'org.tizen.tizen_package_manager_example');