+++ packages/tizen_package_manager/CHANGELOG.md
@@ -1,4 +1,11 @@
+## 0.2.0
+
+* Rename `package_manager.dart` to `tizen_package_manager.dart`.
+* Create an enum type `StorageType` for `PackageInfo.installedStorageType`.
+* Refactor the C++ code.
+* Revise the example app.
+* General cleanups.
+
 ## 0.1.0
 
 * Initial release.
-